+++ apps/nextjs-example/components/AppContext.tsx
@@ -5,6 +5,7 @@ import { PontemWallet } from "@pontem/wallet-adapter-plugin";
 import { TrustWallet } from "@trustwallet/aptos-wallet-adapter";
 import { SpikaWallet } from "@spika/aptos-plugin";
 import { AptosWalletAdapterProvider } from "@aptos-labs/wallet-adapter-react";
+import { FewchaWallet } from "fewcha-plugin-wallet-adapter";
 import {
   AutoConnectProvider,
   useAutoConnect,
@@ -21,6 +22,7 @@ const WalletContextProvider: FC<{ children: ReactNode }> = ({ children }) => {
     new PontemWallet(),
     new TrustWallet(),
     new SpikaWallet(),
+    new FewchaWallet(),
   ];
 
   return (