+++ apps/nextjs-example/pages/index.tsx
@@ -1,6 +1,7 @@
 import { AptosClient, Types } from "aptos";
 import { useWallet } from "@aptos-labs/wallet-adapter-react";
 import { Fragment, useState } from "react";
+import { WalletSelector } from "@aptos-labs/wallet-adapter-ant-design";
 import { ErrorAlert, SuccessAlert } from "../components/Alert";
 import dynamic from "next/dynamic";
 import Image from "next/image";
@@ -66,6 +67,7 @@ export default function App() {
     };
     try {
       const response = await signTransaction(payload);
+      setSuccessAlertMessage(JSON.stringify(response));
       console.log("response", response);
     } catch (error: any) {
       console.log("error", error);
@@ -124,12 +126,20 @@ export default function App() {
               <WalletButtons />
             </td>
           </tr>
-          <tr>
+                    <tr>
             <td className="px-8 py-4 w-1/4">
               <h3>Tailwind Wallet Adapter</h3>
             </td>
             <td className="px-8 py-4 w-3/4">
               <WalletDropdown />
+            </td>
+          <tr>
+          <tr>
+            <td className="px-8 border-t py-4 w-1/4">
+              <h3>Wallet Selector</h3>
+            </td>
+            <td className="px-8 py-4 border-t w-3/4">
+              <WalletSelector />
             </td>
           </tr>
           <tr>