+++ apps/nextjs-example/tailwind.config.js
@@ -7,4 +7,7 @@ module.exports = {
   theme: {
   },
   plugins: [],
+  corePlugins: {
+    preflight: false,
+  },
 };