+++ apps/nextjs-example/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @aptos-labs/wallet-adapter-nextjs-example
 
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies [552d255]
+  - @aptos-labs/wallet-adapter-react@0.2.2
+
+## 0.2.0
+
+### Minor Changes
+
+- 8124d54: Create wallet selector modal with ant design
+
+### Patch Changes
+
+- Updated dependencies [8124d54]
+  - @aptos-labs/wallet-adapter-ant-design@0.1.0
+
+## 0.1.1
+
+### Patch Changes
+
+- c3eb031: Export WalletReadyState enum from react package
+- Updated dependencies [c3eb031]
+  - @aptos-labs/wallet-adapter-react@0.2.1
+
 ## 0.1.0
 
 ### Minor Changes